Display only matching composer presets as New Topic dropdown options
We already reached to 18 composer presets on our site, and that makes selecting a the New Topic button dropdown option harder for the users.

Even though, we inject the word “Recommended” for the options with `is-selected` class, seeing the long list is confusing and overwhelming for some users.

This PR adds a couple new settings:
 * `hide_unmatched_composer_presets`: Hide the composer presets that are NOT highlighted as recommendation.
* `hide_new_topic_button_on_default_page`: "Hide 'New Topic' button on default page. Uncheck to display the button on all pages.